Rename demo components
