+++ docs/data/data-grid/demo/PopularFeaturesDemo.js
@@ -1,9 +1,8 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import {
   DataGridPremium,
   gridClasses,
-  GridColDef,
-  GridRenderCellParams,
   GridToolbarContainer,
   GridToolbarQuickFilter,
 } from '@mui/x-data-grid-premium';
@@ -11,24 +10,26 @@ import Link from '@mui/material/Link';
 import Chip from '@mui/material/Chip';
 import Avatar from '@mui/material/Avatar';
 import Typography from '@mui/material/Typography';
-import DemoHub, { featuresSet } from './DemoHub';
+import { featuresSet } from './features';
+import Box from '@mui/material/Box';
 
-const PlanTag = (props: { plan: string }) => {
-  function getChipProperties(plan: string) {
-    switch (plan.toLowerCase()) {
-      case 'premium':
-        return { avatarLink: '/static/x/premium.svg', color: '#ffecc8' };
-      case 'pro':
-        return { avatarLink: '/static/x/pro.svg', color: '#c8e9ff' };
-      default:
-        return { avatarLink: undefined, color: '#c8ffdb' };
-    }
+function getChipProperties(plan) {
+  switch (plan.toLowerCase()) {
+    case 'premium':
+      return { avatarLink: '/static/x/premium.svg', color: '#ffecc8' };
+    case 'pro':
+      return { avatarLink: '/static/x/pro.svg', color: '#c8e9ff' };
+    default:
+      return { avatarLink: undefined, color: '#c8ffdb' };
   }
+}
 
+const PlanTag = (props) => {
   const chipPropperties = getChipProperties(props.plan);
   const avatar = !chipPropperties.avatarLink ? undefined : (
     <Avatar src={chipPropperties.avatarLink} />
   );
+
   return (
     <Chip
       avatar={avatar}
@@ -38,8 +39,24 @@ const PlanTag = (props: { plan: string }) => {
   );
 };
 
+PlanTag.propTypes = {
+  plan: PropTypes.string.isRequired,
+};
+
+function CustomToolbar() {
+  return (
+    <GridToolbarContainer sx={{ p: 1 }}>
+      <GridToolbarQuickFilter />
+    </GridToolbarContainer>
+  );
+}
+
+const renderFeatures = (row) => {
+  return <Box sx={{ width: '80%', margin: 'auto', py: 2 }}>{row.demo}</Box>;
+};
+
 function OneMasterDemo() {
-  const columns: GridColDef[] = [
+  const columns = [
     {
       field: 'name',
       headerName: 'Feature name',
@@ -73,7 +90,7 @@ function OneMasterDemo() {
       width: 130,
       type: 'singleSelect',
       valueOptions: ['Premium', 'Pro', 'Community'],
-      renderCell: (params: GridRenderCellParams<string>) => {
+      renderCell: (params) => {
         if (!params.value) {
           return <React.Fragment />;
         }
@@ -82,14 +99,6 @@ function OneMasterDemo() {
     },
   ];
 
-  function CustomToolbar() {
-    return (
-      <GridToolbarContainer sx={{ p: 1 }}>
-        <GridToolbarQuickFilter />
-      </GridToolbarContainer>
-    );
-  }
-
   return (
     <div style={{ height: 600, width: '100%' }}>
       <DataGridPremium
@@ -97,7 +106,7 @@ function OneMasterDemo() {
         componentsProps={{
           toolbar: { showQuickFilter: true, quickFilterProps: { debounceMs: 500 } },
         }}
-        getDetailPanelContent={({ row }) => DemoHub(row)}
+        getDetailPanelContent={({ row }) => renderFeatures(row)}
         getDetailPanelHeight={() => 'auto'}
         getRowHeight={() => 'auto'}
         sx={{
