+++ docs/data/data-grid/demo/features.tsx
@@ -1,12 +1,11 @@
 import React from 'react';
-import { Box } from '@mui/material';
 import EditingWithDatePickers from '../editing/EditingWithDatePickers';
 import ExcelExport from '../export/ExcelExport';
 import QuickFilteringGrid from '../filtering/QuickFilteringGrid';
 import BasicDetailPanels from '../master-detail/BasicDetailPanels';
 import RowGroupingInitialState from '../row-grouping/RowGroupingInitialState';
 import FullFeaturedDemo from './FullFeaturedDemo';
-import ControlPinnedColumns from '../column-pinning/ControlPinnedColumns';
+import BasicColumnPinning from '../column-pinning/BasicColumnPinning';
 import ColumnVirtualizationGrid from '../virtualization/ColumnVirtualizationGrid';
 import CursorPaginationGrid from '../pagination/CursorPaginationGrid';
 import ColumnSelectorGrid from '../column-visibility/ColumnSelectorGrid';
@@ -25,8 +24,8 @@ export const featuresSet = [
   },
   {
     id: 2,
-    name: 'Editing inline',
-    description: 'Edit your data inside the grid',
+    name: 'Inline editing',
+    description: 'Edit your data inside the cells by double-clicking or pressing Enter',
     plan: 'Community',
     detailPage: '/editing/',
     demo: <EditingWithDatePickers />,
@@ -41,9 +40,9 @@ export const featuresSet = [
   },
   {
     id: 4,
-    name: 'Save and Restore state',
+    name: 'Save and restore state',
     description:
-      'Save and Restore internal state and configurations like active filters and sorting',
+      'Save and restore internal state and configurations like active filters and sorting',
     plan: 'Community',
     detailPage: '/state/#save-and-restore-the-state',
     demo: <RestoreStateInitialState />,
@@ -87,7 +86,7 @@ export const featuresSet = [
     description: 'Pin your columns to the left or right',
     plan: 'Pro',
     detailPage: '/column-pinning/',
-    demo: <ControlPinnedColumns />,
+    demo: <BasicColumnPinning />,
   },
   {
     id: 10,
@@ -124,8 +123,3 @@ export const featuresSet = [
   },
 ];
 
-const DemoHub = (row) => {
-  return <Box sx={{ width: '80%', margin: 'auto', py: 2 }}>{row.demo}</Box>;
-};
-
-export default DemoHub;
